backend/warehouse-reminder/requirements.txt
psycopg2-binary==2.9.9
aiogram==3.4.1
